requirements.txt
requests
obsws-python
psutil
pyinstaller
Pillow
# tkinter (included with standard Python on Windows) 